+++ core/config_manager/ConfigManagerBase.cpp
@@ -308,11 +308,11 @@ void ConfigManagerBase::UpdatePluginStats(const Json::Value& config) {
                     std::string type = config["plugin"][*it][i]["type"].asString();
                     stats[*it].insert(type);
                     if (type == "service_docker_stdout") {
-                        if (config["plugin"][*it][i].isMember("detail") && config["plugin"][*it][i]["detail"].isObject() 
-                            && config["plugin"][*it][i]["detail"].isMember("CollectContainersFlag") 
+                        if (config["plugin"][*it][i].isMember("detail") && config["plugin"][*it][i]["detail"].isObject()
+                            && config["plugin"][*it][i]["detail"].isMember("CollectContainersFlag")
                             && config["plugin"][*it][i]["detail"]["CollectContainersFlag"].isBool()
                             && config["plugin"][*it][i]["detail"]["CollectContainersFlag"].asBool()) {
-                                stats["inner_function"].insert("collect_containers_meta");
+                            stats["inner_function"].insert("collect_containers_meta");
                         }
                     }
                 }
@@ -341,8 +341,10 @@ void ConfigManagerBase::UpdatePluginStats(const Json::Value& config) {
         stats["processors"].insert(processor);
         stats["flushers"].insert("flusher_sls");
 
-        if (config.isMember("advanced") && config["advanced"].isObject() && config["advanced"].isMember("collect_containers_flag") 
-            && config["advanced"]["collect_containers_flag"].isBool() && config["advanced"]["collect_containers_flag"].asBool()) {
+        if (config.isMember("advanced") && config["advanced"].isObject()
+            && config["advanced"].isMember("collect_containers_flag")
+            && config["advanced"]["collect_containers_flag"].isBool()
+            && config["advanced"]["collect_containers_flag"].asBool()) {
             stats["inner_function"].insert("collect_containers_meta");
         }
     }
@@ -510,13 +512,12 @@ void ConfigManagerBase::LoadSingleUserConfig(const std::string& logName, const J
 
             if (logType == PLUGIN_LOG) {
                 config = new Config(
-                    "", "", logType, logName, "", projectName, false, 0, 0, category, false, "", discardUnmatch);
+                    "", "", logType, logName, "", "", "", projectName, false, 0, 0, category, false, "", discardUnmatch);
                 if (pluginConfig.empty()) {
                     throw ExceptionBase(std::string("The plugin log type is invalid"));
                 }
                 if (!pluginConfigJson.isNull()) {
-                    pluginConfigJson = ConfigManager::GetInstance()->CheckPluginProcessor(pluginConfigJson, value);
-                    pluginConfig = ConfigManager::GetInstance()->CheckPluginFlusher(pluginConfigJson);
+                    config->mPluginProcessFlag = true;
                     if (pluginConfig.find("\"observer_ilogtail_") != string::npos) {
                         if (pluginConfigJson.isMember("inputs")) {
                             if (pluginConfigJson["inputs"].isObject() || pluginConfigJson["inputs"].isArray()) {
@@ -527,25 +528,26 @@ void ConfigManagerBase::LoadSingleUserConfig(const std::string& logName, const J
                             if (pluginConfigJson.isMember("processors")
                                 && (pluginConfigJson["processors"].isObject()
                                     || pluginConfigJson["processors"].isArray())) {
-                                config->mPluginProcessFlag = true;
                                 SetNotFoundJsonMember(pluginConfigJson, MIX_PROCESS_MODE, "observer");
-                                config->mPluginConfig = pluginConfigJson.toStyledString();
                             }
-
                         } else {
                             LOG_WARNING(sLogger,
                                         ("observer config is not a legal JSON object",
                                          logName)("project", projectName)("logstore", category));
+                            throw ExceptionBase(std::string("observer config is not a legal JSON object"));
                         }
-                    } else {
-                        config->mPluginConfig = pluginConfig;
                     }
+                    pluginConfigJson = ConfigManager::GetInstance()->CheckPluginProcessor(pluginConfigJson, value);
+                    pluginConfig = ConfigManager::GetInstance()->CheckPluginFlusher(pluginConfigJson);
+                    config->mPluginConfig = pluginConfig;
                 }
             } else if (logType == STREAM_LOG) {
                 config = new Config("",
                                     "",
                                     logType,
                                     logName,
+                                    "",
+                                    "",
                                     "",
                                     projectName,
                                     false,
@@ -576,10 +578,22 @@ void ConfigManagerBase::LoadSingleUserConfig(const std::string& logName, const J
                 if (size > 0 && PATH_SEPARATOR[0] == logPath[size - 1])
                     logPath = logPath.substr(0, size - 1);
 
-                string logBeingReg = GetStringValue(value, "log_begin_reg", "");
-                if (logBeingReg != "" && CheckRegFormat(logBeingReg) == false) {
-                    throw ExceptionBase("The log begin line is not value regex : " + logBeingReg);
+                string logBeginReg = GetStringValue(value, "log_begin_reg", "");
+                if (logBeginReg != "" && CheckRegFormat(logBeginReg) == false) {
+                    throw ExceptionBase("The log begin line is not value regex : " + logBeginReg);
                 }
+                string logContinueReg = GetStringValue(value, "log_continue_reg", "");
+                if (logContinueReg != "" && CheckRegFormat(logContinueReg) == false) {
+                    throw ExceptionBase("The log continue line is not value regex : " + logContinueReg);
+                }
+                string logEndReg = GetStringValue(value, "log_end_reg", "");
+                if (logEndReg != "" && CheckRegFormat(logEndReg) == false) {
+                    throw ExceptionBase("The log end line is not value regex : " + logEndReg);
+                }
+                int readerFlushTimeout = 5;
+                if (value.isMember("reader_flush_timeout"))
+                    readerFlushTimeout = GetIntValue(value, "reader_flush_timeout");
+
                 string filePattern = GetStringValue(value, "file_pattern");
                 // raw log flag
                 bool rawLogFlag = false;
@@ -591,7 +605,9 @@ void ConfigManagerBase::LoadSingleUserConfig(const std::string& logName, const J
                                     filePattern,
                                     logType,
                                     logName,
-                                    logBeingReg,
+                                    logBeginReg,
+                                    logContinueReg,
+                                    logEndReg,
                                     projectName,
                                     isPreserve,
                                     preserveDepth,
@@ -599,20 +615,38 @@ void ConfigManagerBase::LoadSingleUserConfig(const std::string& logName, const J
                                     category,
                                     rawLogFlag,
                                     "",
-                                    discardUnmatch);
+                                    discardUnmatch,
+                                    readerFlushTimeout);
 
                 // normal log file config can have plugin too
+                // Boolean force_enable_pipeline.
+                if (value.isMember("force_enable_pipeline") && value["force_enable_pipeline"].isBool()
+                    && value["force_enable_pipeline"].asBool()) {
+                    config->mForceEnablePipeline = true;
+                    LOG_INFO(sLogger,
+                             ("set force enable pipeline",
+                              config->mForceEnablePipeline)("project", projectName)("config", logName));
+                }
                 if (!pluginConfig.empty() && !pluginConfigJson.isNull()) {
+                    if (pluginConfigJson.isMember("processors")
+                        && (pluginConfigJson["processors"].isObject() || pluginConfigJson["processors"].isArray())
+                        && !pluginConfigJson["processors"].empty()) {
+                        config->mPluginProcessFlag = true;
+                    }
+                    if (pluginConfigJson.isMember("flushers")
+                        && (pluginConfigJson["flushers"].isObject() || pluginConfigJson["flushers"].isArray())
+                        && !pluginConfigJson["flushers"].empty() && IsMeaningfulFlusher(pluginConfigJson["flushers"])) {
+                        config->mPluginProcessFlag = true;
+                    }
                     // check processors
                     // set process flag when config have processors
                     if (pluginConfigJson.isMember("processors")
                         && (pluginConfigJson["processors"].isObject() || pluginConfigJson["processors"].isArray())) {
                         // patch enable_log_position_meta to split processor if exists ...
-                        config->mPluginProcessFlag = true;
                         pluginConfigJson = ConfigManager::GetInstance()->CheckPluginProcessor(pluginConfigJson, value);
                         pluginConfig = ConfigManager::GetInstance()->CheckPluginFlusher(pluginConfigJson);
-                        config->mPluginConfig = pluginConfig;
                     }
+                    config->mPluginConfig = pluginConfig;
                 }
                 if (value.isMember("docker_file") && value["docker_file"].isBool() && value["docker_file"].asBool()) {
                     if (AppConfig::GetInstance()->IsPurageContainerMode()) {
@@ -1088,6 +1122,21 @@ LogFilterRule* ConfigManagerBase::GetFilterFule(const Json::Value& filterKeys, c
         }
     }
     return rulePtr;
+}
+
+bool ConfigManagerBase::IsMeaningfulFlusher(const Json::Value& flusherConfig) {
+    // Only one sls flusher without any option is not meaningful to use go plugin
+    if (flusherConfig.isArray() && flusherConfig.size() == 1) {
+        const auto& flusher = flusherConfig[0];
+        try {
+            if (flusher["type"].asString() == "flusher_sls" && flusher["detail"].empty()) {
+                return false;
+            }
+        } catch (Json::Exception& e) {
+            LOG_WARNING(sLogger, ("parse flusher plugin json failed", e.what()));
+        }
+    }
+    return true;
 }
 
 bool ConfigManagerBase::LoadAllConfig() {