+++ core/reader/JsonLogFileReader.cpp
@@ -61,13 +61,13 @@ void JsonLogFileReader::SetTimeKey(const std::string& timeKey) {
         mUseSystemTime = false;
 }
 
-bool JsonLogFileReader::ParseLogLine(const char* buffer,
+bool JsonLogFileReader::ParseLogLine(StringView buffer,
                                      sls_logs::LogGroup& logGroup,
                                      ParseLogError& error,
                                      LogtailTime& lastLogLineTime,
                                      std::string& lastLogTimeStr,
                                      uint32_t& logGroupSize) {
-    if (strlen(buffer) == 0)
+    if (buffer.empty())
         return true;
 
     if (logGroup.logs_size() == 0) {
@@ -77,25 +77,28 @@ bool JsonLogFileReader::ParseLogLine(const char* buffer,
     bool parseSuccess = true;
     uint64_t preciseTimestamp = 0;
     rapidjson::Document doc;
-    doc.Parse(buffer);
+    doc.Parse(buffer.data(), buffer.size());
     if (doc.HasParseError()) {
         if (LogtailAlarm::GetInstance()->IsLowLevelAlarmValid()) {
             LOG_WARNING(sLogger,
                         ("parse json log fail, log",
                          buffer)("rapidjson offset", doc.GetErrorOffset())("rapidjson error", doc.GetParseError())(
                             "project", mProjectName)("logstore", mCategory)("file", mHostLogPath));
-            LogtailAlarm::GetInstance()->SendAlarm(
-                PARSE_LOG_FAIL_ALARM, string("parse json fail:") + string(buffer), mProjectName, mCategory, mRegion);
+            LogtailAlarm::GetInstance()->SendAlarm(PARSE_LOG_FAIL_ALARM,
+                                                   string("parse json fail:") + buffer.to_string(),
+                                                   mProjectName,
+                                                   mCategory,
+                                                   mRegion);
         }
         error = PARSE_LOG_FORMAT_ERROR;
         parseSuccess = false;
     } else if (!doc.IsObject()) {
         if (LogtailAlarm::GetInstance()->IsLowLevelAlarmValid()) {
-            LOG_WARNING(
-                sLogger,
-                ("invalid json object, log", buffer)("project", mProjectName)("logstore", mCategory)("file", mHostLogPath));
+            LOG_WARNING(sLogger,
+                        ("invalid json object, log", buffer)("project", mProjectName)("logstore",
+                                                                                      mCategory)("file", mHostLogPath));
             LogtailAlarm::GetInstance()->SendAlarm(PARSE_LOG_FAIL_ALARM,
-                                                   string("invalid json object:") + string(buffer),
+                                                   string("invalid json object:") + buffer.to_string(),
                                                    mProjectName,
                                                    mCategory,
                                                    mRegion);
@@ -105,7 +108,7 @@ bool JsonLogFileReader::ParseLogLine(const char* buffer,
     } else if (!mUseSystemTime) {
         rapidjson::Value::ConstMemberIterator itr = doc.FindMember(mTimeKey.c_str());
         if (itr != doc.MemberEnd() && (itr->value.IsString() || itr->value.IsInt64())) {
-            if (!LogParser::ParseLogTime(buffer,
+            if (!LogParser::ParseLogTime(buffer.data(),
                                          lastLogTimeStr,
                                          lastLogLineTime,
                                          preciseTimestamp,
@@ -131,7 +134,7 @@ bool JsonLogFileReader::ParseLogLine(const char* buffer,
                                 "logstore", mCategory)("file", mHostLogPath));
                 LogtailAlarm::GetInstance()->SendAlarm(PARSE_LOG_FAIL_ALARM,
                                                        string("found no time_key: ") + mTimeKey
-                                                           + ", log:" + string(buffer),
+                                                           + ", log:" + buffer.to_string(),
                                                        mProjectName,
                                                        mCategory,
                                                        mRegion);
@@ -191,18 +194,22 @@ std::string JsonLogFileReader::RapidjsonValueToString(const rapidjson::Value& va
     }
 }
 
-vector<int32_t> JsonLogFileReader::LogSplit(char* buffer, int32_t size, int32_t& lineFeed) {
-    vector<int32_t> index;
-    int32_t i = 0;
-    index.push_back(i);
+bool JsonLogFileReader::LogSplit(const char* buffer,
+                                 int32_t size,
+                                 int32_t& lineFeed,
+                                 std::vector<StringView>& logIndex,
+                                 std::vector<StringView>& discardIndex) {
+    int32_t line_beg = 0, i = 0;
     while (++i < size) {
-        if (buffer[i] == '\0' && i < size - 1)
-            index.push_back(i + 1);
+        if (buffer[i] == '\0') {
+            logIndex.emplace_back(buffer + line_beg, i - line_beg);
+            line_beg = i + 1;
+        }
     }
-    return index;
+    logIndex.emplace_back(buffer + line_beg, size - line_beg);
+    return true;
 }
 
-// rollbackLineFeedCount is useful in encoding conversion, ingnore it in json reader(should be utf8)
 // split log & find last match in this function
 int32_t JsonLogFileReader::LastMatchedLine(char* buffer, int32_t size, int32_t& rollbackLineFeedCount) {
     int32_t readBytes = 0;
@@ -216,8 +223,8 @@ int32_t JsonLogFileReader::LastMatchedLine(char* buffer, int32_t size, int32_t& 
         if (FindJsonMatch(buffer, beginIdx, size, endIdx, startWithBlock)) {
             noJsonBlockFlag = false;
             buffer[endIdx] = '\0';
-            readBytes = endIdx + 1;
             beginIdx = endIdx + 1;
+            readBytes = endIdx + 1;
             rollbackLineFeedCount = 0;
         } else {
             char* pos = strchr(buffer + beginIdx, '\n');
@@ -228,10 +235,11 @@ int32_t JsonLogFileReader::LastMatchedLine(char* buffer, int32_t size, int32_t& 
             buffer[endIdx] = '\0'; // invalid json line
             beginIdx = endIdx + 1;
             if (noJsonBlockFlag) {
-                // if no json block in this buffer and now line has no block, forse read this line
+                // if no json block in this buffer and now line has no block, force read this line
                 // if no json block in this buffer and now line has block, set noJsonBlockFlag false
                 if (!startWithBlock) {
                     readBytes = endIdx + 1;
+                    rollbackLineFeedCount = 0;
                 } else {
                     noJsonBlockFlag = false;
                 }
@@ -259,8 +267,8 @@ bool JsonLogFileReader::FindJsonMatch(
     }
     if (idx == size || buffer[idx] != '{') {
         LOG_DEBUG(sLogger,
-                  ("invalid json begining", buffer + beginIdx)("project", mProjectName)("logstore",
-                                                                                        mCategory)("file", mHostLogPath));
+                  ("invalid json begining",
+                   buffer + beginIdx)("project", mProjectName)("logstore", mCategory)("file", mHostLogPath));
         startWithBlock = false;
         return false;
     }
@@ -302,7 +310,7 @@ bool JsonLogFileReader::FindJsonMatch(
         }
     }
     LOG_DEBUG(sLogger,
-              ("find no match, beginIdx", beginIdx)("idx", idx)("project", mProjectName)("logstore",
-                                                                                         mCategory)("file", mHostLogPath));
+              ("find no match, beginIdx",
+               beginIdx)("idx", idx)("project", mProjectName)("logstore", mCategory)("file", mHostLogPath));
     return false;
 }