+++ plugin_main/plugin_export.go
@@ -149,10 +149,20 @@ func ProcessRawLogV2(configName string, rawLog []byte, packID string, topic stri
 func ProcessLog(configName string, logBytes []byte, packID string, topic string, tags []byte) int {
 	config, exists := pluginmanager.LogtailConfig[configName]
 	if !exists {
-		logger.Debug(context.Background(), "ProcessLog not found", configName)
+		logger.Debug(context.Background(), "config not found", configName)
 		return -1
 	}
 	return config.ProcessLog(logBytes, packID, util.StringDeepCopy(topic), tags)
+}
+
+//export ProcessLogGroup
+func ProcessLogGroup(configName string, logBytes []byte, packID string) int {
+	config, exists := pluginmanager.LogtailConfig[configName]
+	if !exists {
+		logger.Debug(context.Background(), "config not found", configName)
+		return -1
+	}
+	return config.ProcessLogGroup(logBytes, packID)
 }
 
 //export HoldOn