+++ test/case/core/reader_new_line_after_timeout/run.sh
@@ -1,0 +1,19 @@
+#!/usr/bin/env bash
+
+# Copyright 2021 iLogtail Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+sleep 5
+echo -n "5=====" >> /root/a.log  # create an event to trigger read
+sleep 3
+echo -e "\n6=====" >> /root/a.log